third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/slot-element-focusable.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] slot element with default style should be focusable
  assert_equals: slot is now focused (style) expected "slot-is-focused" but got "slot-not-focused"
Harness: the test ran to completion.
